requirements.txt
faster-whisper==1.0.2
groq==0.6.0
openai==1.30.1
google-generativeai==0.5.4
opencv-python==4.9.0.80
pillow==10.3.0
Pyaudio==0.2.14
SpeechRecognition==3.10.1
pyperclip==1.8.2
python-dotenv